Application Web/car-price-maroc/requirements.txt
pycaret==3.0.0rc4
catboost==1.1
bokeh==2.4.3
bokeh_catplot==0.1.9
seaborn==0.11.2
streamlit==1.13.0
streamlit-shap==1.0.2
streamlit-aggrid==0.3.2
explainerdashboard==0.4.0
dash-bootstrap-components==1.2.1
evidently==0.1.59.dev0
streamlit-pandas-profiling==0.1.3
pandas-profiling==3.4.0
